requirement.txt
alembic==1.16.2
blinker==1.9.0
Brotli==1.1.0
cffi==1.17.1
click==8.1.8
cssselect2==0.8.0
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.59.0
greenlet==3.2.3
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
pillow==11.3.0
pycparser==2.22
pydyf==0.11.0
PyMySQL==1.1.1
pyphen==0.17.2
python-dotenv==1.1.1
SQLAlchemy==2.0.41
tinycss2==1.4.0
tinyhtml5==2.0.0
tomli==2.2.1
typing_extensions==4.14.0
weasyprint==66.0
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.2.1
zipp==3.23.0
zopfli==0.2.3.post1
